# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Boiler Service Technician API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List properties
      parameters:
        - name: customer_id
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Property list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Property details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyDetail"
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePropertyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Soft-delete property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
...
  /service-records:
    post:
      operationId: createServiceRecord
      tags: [serviceRecords]
      summary: Create a service record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRecordBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRecord"
  /service-records/{id}:
    get:
      operationId: getServiceRecord
      tags: [serviceRecords]
      summary: Get service record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Service record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRecord"
    patch:
      operationId: updateServiceRecord
      tags: [serviceRecords]
      summary: Update service record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRecordBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRecord"
  /service-records/job/{jobId}:
    get:
      operationId: getServiceRecordByJob
      tags: [serviceRecords]
      summary: Get service record by job ID
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Service record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRecord"
...
  /signatures:
    post:
      operationId: createSignature
      tags: [signatures]
      summary: Upload a signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSignatureBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signature"
  /signatures/job/{jobId}:
    get:
      operationId: getJobSignatures
      tags: [signatures]
      summary: Get signatures for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Signatures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signature"
...
  /reports/completed-by-technician:
    get:
      operationId: getCompletedByTechnician
      tags: [reports]
      summary: Get completed jobs grouped by technician
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Completed jobs by technician
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TechnicianStats"
components:
  schemas:
...
    Customer:
      type: object
      required: [id, first_name, last_name, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: ["string", "null"]
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        address_line1:
          type: ["string", "null"]
        address_line2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        postcode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Property:
      type: object
      required: [id, customer_id, address_line1, postcode, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        address_line1:
          type: string
        address_line2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        postcode:
          type: string
        property_type:
          type: ["string", "null"]
        occupancy_type:
          type: ["string", "null"]
        access_notes:
          type: ["string", "null"]
        boiler_location:
          type: ["string", "null"]
        flue_location:
          type: ["string", "null"]
        tank_location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        customer_name:
          type: ["string", "null"]
    PropertyDetail:
      allOf:
        - $ref: "#/components/schemas/Property"
        - type: object
          properties:
            customer:
              $ref: "#/components/schemas/Customer"
            appliances:
              type: array
              items:
                $ref: "#/components/schemas/Appliance"
            recent_jobs:
              type: array
              items:
                $ref: "#/components/schemas/JobListItem"
    CreatePropertyBody:
      type: object
      required: [customer_id, address_line1, postcode]
      properties:
        customer_id:
          type: string
          format: uuid
        address_line1:
          type: string
        address_line2:
          type: string
        city:
          type: string
        county:
          type: string
        postcode:
          type: string
        property_type:
          type: string
        occupancy_type:
          type: string
        access_notes:
          type: string
        boiler_location:
          type: string
        flue_location:
          type: string
        tank_location:
          type: string
        notes:
          type: string
    UpdatePropertyBody:
      type: object
      properties:
        customer_id:
          type: string
          format: uuid
        address_line1:
          type: string
        address_line2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        postcode:
          type: string
        property_type:
          type: ["string", "null"]
        occupancy_type:
          type: ["string", "null"]
        access_notes:
          type: ["string", "null"]
        boiler_location:
          type: ["string", "null"]
        flue_location:
          type: ["string", "null"]
        tank_location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
    Appliance:
      type: object
      required: [id, property_id, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        manufacturer:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serial_number:
          type: ["string", "null"]
        boiler_type:
          type: ["string", "null"]
        fuel_type:
          type: ["string", "null"]
        system_type:
          type: ["string", "null"]
        installation_date:
          type: ["string", "null"]
          format: date
        warranty_expiry:
          type: ["string", "null"]
          format: date
        burner_make:
          type: ["string", "null"]
        burner_model:
          type: ["string", "null"]
        nozzle_size:
          type: ["string", "null"]
        pump_pressure:
          type: ["string", "null"]
        controls:
          type: ["string", "null"]
        last_service_date:
          type: ["string", "null"]
          format: date
        next_service_due:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        property_address:
          type: ["string", "null"]
...
    JobListItem:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            customer_name:
              type: ["string", "null"]
            property_address:
              type: ["string", "null"]
            technician_name:
              type: ["string", "null"]
...
    ServiceRecord:
      type: object
      required: [id, job_id, technician_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
        arrival_time:
          type: ["string", "null"]
          format: date-time
        departure_time:
          type: ["string", "null"]
          format: date-time
        visual_inspection:
          type: ["string", "null"]
        appliance_condition:
          type: ["string", "null"]
        flue_inspection:
          type: ["string", "null"]
        combustion_co2:
          type: ["string", "null"]
        combustion_co:
          type: ["string", "null"]
        combustion_o2:
          type: ["string", "null"]
        combustion_temp:
          type: ["string", "null"]
        combustion_efficiency:
          type: ["string", "null"]
        smoke_test:
          type: ["string", "null"]
        smoke_number:
          type: ["string", "null"]
        burner_cleaned:
          type: boolean
        heat_exchanger_cleaned:
          type: boolean
        nozzle_checked:
          type: boolean
        nozzle_replaced:
          type: boolean
        nozzle_size_fitted:
          type: ["string", "null"]
        electrodes_checked:
          type: boolean
        electrodes_replaced:
          type: boolean
        filter_checked:
          type: boolean
        filter_cleaned:
          type: boolean
        filter_replaced:
          type: boolean
        oil_line_checked:
          type: boolean
        fire_valve_checked:
          type: boolean
        seals_gaskets_checked:
          type: boolean
        seals_gaskets_replaced:
          type: boolean
        controls_checked:
          type: boolean
        thermostat_checked:
          type: boolean
        safety_devices_checked:
          type: boolean
        safety_devices_notes:
          type: ["string", "null"]
        leaks_found:
          type: boolean
        leaks_details:
          type: ["string", "null"]
        defects_found:
          type: boolean
        defects_details:
          type: ["string", "null"]
        advisories:
          type: ["string", "null"]
        parts_required:
          type: ["string", "null"]
        work_completed:
          type: ["string", "null"]
        appliance_safe:
          type: boolean
        follow_up_required:
          type: boolean
        follow_up_notes:
          type: ["string", "null"]
        next_service_due:
          type: ["string", "null"]
          format: date
        additional_notes:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateServiceRecordBody:
      type: object
      required: [job_id, technician_id]
      properties:
        job_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
        arrival_time:
          type: string
          format: date-time
        departure_time:
          type: string
          format: date-time
        visual_inspection:
          type: string
        appliance_condition:
          type: string
        flue_inspection:
          type: string
        combustion_co2:
          type: string
        combustion_co:
          type: string
        combustion_o2:
          type: string
        combustion_temp:
          type: string
        combustion_efficiency:
          type: string
        smoke_test:
          type: string
        smoke_number:
          type: string
        burner_cleaned:
          type: boolean
        heat_exchanger_cleaned:
          type: boolean
        nozzle_checked:
          type: boolean
        nozzle_replaced:
          type: boolean
        nozzle_size_fitted:
          type: string
        electrodes_checked:
          type: boolean
        electrodes_replaced:
          type: boolean
        filter_checked:
          type: boolean
        filter_cleaned:
          type: boolean
        filter_replaced:
          type: boolean
        oil_line_checked:
          type: boolean
        fire_valve_checked:
          type: boolean
        seals_gaskets_checked:
          type: boolean
        seals_gaskets_replaced:
          type: boolean
        controls_checked:
          type: boolean
        thermostat_checked:
          type: boolean
        safety_devices_checked:
          type: boolean
        safety_devices_notes:
          type: string
        leaks_found:
          type: boolean
        leaks_details:
          type: string
        defects_found:
          type: boolean
        defects_details:
          type: string
        advisories:
          type: string
        parts_required:
          type: string
        work_completed:
          type: string
        appliance_safe:
          type: boolean
        follow_up_required:
          type: boolean
        follow_up_notes:
          type: string
        next_service_due:
          type: string
          format: date
        additional_notes:
          type: string
    UpdateServiceRecordBody:
      type: object
      properties:
        arrival_time:
          type: ["string", "null"]
          format: date-time
        departure_time:
          type: ["string", "null"]
          format: date-time
        visual_inspection:
          type: ["string", "null"]
        appliance_condition:
          type: ["string", "null"]
        flue_inspection:
          type: ["string", "null"]
        combustion_co2:
          type: ["string", "null"]
        combustion_co:
          type: ["string", "null"]
        combustion_o2:
          type: ["string", "null"]
        combustion_temp:
          type: ["string", "null"]
        combustion_efficiency:
          type: ["string", "null"]
        smoke_test:
          type: ["string", "null"]
        smoke_number:
          type: ["string", "null"]
        burner_cleaned:
          type: boolean
        heat_exchanger_cleaned:
          type: boolean
        nozzle_checked:
          type: boolean
        nozzle_replaced:
          type: boolean
        nozzle_size_fitted:
          type: ["string", "null"]
        electrodes_checked:
          type: boolean
        electrodes_replaced:
          type: boolean
        filter_checked:
          type: boolean
        filter_cleaned:
          type: boolean
        filter_replaced:
          type: boolean
        oil_line_checked:
          type: boolean
        fire_valve_checked:
          type: boolean
        seals_gaskets_checked:
          type: boolean
        seals_gaskets_replaced:
          type: boolean
        controls_checked:
          type: boolean
        thermostat_checked:
          type: boolean
        safety_devices_checked:
          type: boolean
        safety_devices_notes:
          type: ["string", "null"]
        leaks_found:
          type: boolean
        leaks_details:
          type: ["string", "null"]
        defects_found:
          type: boolean
        defects_details:
          type: ["string", "null"]
        advisories:
          type: ["string", "null"]
        parts_required:
          type: ["string", "null"]
        work_completed:
          type: ["string", "null"]
        appliance_safe:
          type: boolean
        follow_up_required:
          type: boolean
        follow_up_notes:
          type: ["string", "null"]
        next_service_due:
          type: ["string", "null"]
          format: date
        additional_notes:
          type: ["string", "null"]
...
    Signature:
      type: object
      required: [id, job_id, signer_type, signer_name, storage_path, signed_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        signer_type:
          type: string
          enum: [technician, customer]
        signer_name:
          type: string
        storage_path:
          type: string
        signed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        signed_url:
          type: ["string", "null"]
    CreateSignatureBody:
      type: object
      required: [job_id, signer_type, signer_name, image_data]
      properties:
        job_id:
          type: string
          format: uuid
        signer_type:
          type: string
          enum: [technician, customer]
        signer_name:
          type: string
        image_data:
          type: string
...
    TechnicianStats:
      type: object
      required: [technician_id, technician_name, completed_count]
      properties:
        technician_id:
          type: string
          format: uuid
        technician_name:
          type: string
        completed_count:
          type: integer
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/JobListItem"
...